refactor: 시계열 파싱 및 타입 변환 클래스에 역할을 드러내기 위한 네이밍 변경, domain/support로 패키지 이동
